[FIX]: redux setup
