+++ src/helpers/enums.ts
@@ -1,11 +1,3 @@
-/* eslint-disable no-unused-vars */
-
-export enum LoadingStatus {
-  IDLE = 'IDLE',
-  PENDING = 'PENDING',
-  FULLFILLED = 'FULLFILLED',
-  REJECTED = 'REJECTED'
-}
 
 export enum Gender {
   MALE = 'MALE',
