+++ src/store/redux/slices/demo.slice.ts
@@ -1,8 +1,8 @@
 
-import {LoadingStatus} from '@app/helpers';
 import {
   createSlice
 } from '@reduxjs/toolkit';
+import {LoadingStatus} from '../utils';
 
 const SLICE_FEATURE_KEY = 'demo';
